+++ apps/api/src/modules/users/users.controller.ts
@@ -133,7 +133,7 @@ const getMyActivity = asyncHandler(async (req, res) => {
      FROM transactions
      WHERE user_id = $1
      ORDER BY created_at DESC
-     LIMIT $2 OFFSET $3`,
+     LIMIT $2::int OFFSET $3::int`,
     [user.id, limit, offset],
   );
